requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
asyncpg==0.29.0
python-dotenv==1.0.1
python-telegram-bot==20.7
prometheus-client==0.19.0
aioredis==2.0.1
pydantic[email]==2.5.0
structlog==23.2.0